requirement.txt
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
databases==0.9.0
distro==1.9.0
ecdsa==0.19.1
fastapi==0.119.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.7.0
numpy==2.3.3
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pdfminer.six==20250506
pdfplumber==0.11.7
perplexityai==0.16.0
pillow==11.3.0
protobuf==6.32.1
pyarrow==21.0.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
pydeck==0.9.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
watchdog==6.0.0
